moving upper_tri_to_full and optimizing
